pyproject.toml
[project]
name = "hidemyemail-cli"
version = "0.1.0"
description = "CLI tool for managing Apple Hide My Email addresses with Touch ID authentication"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "shellingham>=1.5.0",
    "pyicloud @ git+https://github.com/timlaing/pyicloud.git",
    "pyobjc-framework-LocalAuthentication>=10.0",
    "pyobjc-framework-Security>=10.0",
    "pyobjc-core>=10.0",
]

[project.scripts]
hme = "hidemyemail.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/hidemyemail"]
